# storage/engine: add rocksDBBatchBuilder