@petermattis 2016-05-23
Review status: 0 of 14 files reviewed at latest revision, 1 unresolved discussion.

---

*[storage/replica.go, line 1856 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6861#-KIUIdDQOzTidl34s7_4:-KIUIdDQOzTidl34s7_5:-1787567903) ([raw file](https://github.com/cockroachdb/cockroach/blob/7faa9f846dba58883c29ebb04f83e3366c8900b8/storage/replica.go#L1856)):*
> ```Go
> 		// distinct spans so we don't need the batch to read its own writes.
> 		if b, ok := batch.(engine.Batch); ok {
> 			b.FreezeFlushes()
> ```

I'm not terribly happy with either of the uses of `FreezeFlushes`, but without them we lose the performance benefit here. I'm open to suggestions for how to avoid them or make them safer or some other alternative.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6861)*
<!-- Sent from Reviewable.io -->