@bdarnell 2016-05-24
Review status: 0 of 14 files reviewed at latest revision, 2 unresolved discussions.

---

*[storage/replica.go, line 1666 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6861#-KIW0cTkJZf_EiXUz8_4:-KIW0cTkJZf_EiXUz8_5:1310269292) ([raw file](https://github.com/cockroachdb/cockroach/blob/7faa9f846dba58883c29ebb04f83e3366c8900b8/storage/replica.go#L1666)):*
> ```Go
> 
> 	// Freeze batch flushes as the only remaining use of the batch is for
> 	// range-local keys which we know have not been previously written.
> ```

The applied index key has definitely been written; it's overwritten every time. Where are the reads that you're avoiding with this?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6861)*
<!-- Sent from Reviewable.io -->