@petermattis 2016-05-24
Review status: 0 of 14 files reviewed at latest revision, 2 unresolved discussions.

---

*[storage/replica.go, line 1666 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6861#-KIW0cTkJZf_EiXUz8_4:-KIXiP3SBMoEQGbU_TKj:-215162612) ([raw file](https://github.com/cockroachdb/cockroach/blob/7faa9f846dba58883c29ebb04f83e3366c8900b8/storage/replica.go#L1666)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary>
> The applied index key has definitely been written; it's overwritten every time. Where are the reads that you're avoiding with this?
</details>
The comment might be confusing. I'm trying to avoid a read of data written to the batch which requires flushing the batch from Go down into C++. We're definitely still reading previously written state, just not state that was written within this batch. As I mention below, this API is annoying. It would be better if I could figure out how to make this performant without doing so, but so far I haven't been able to.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6861)*
<!-- Sent from Reviewable.io -->