@petermattis 2016-05-24
Review status: 0 of 14 files reviewed at latest revision, 2 unresolved discussions.

---

*[storage/replica.go, line 1666 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6861#-KIW0cTkJZf_EiXUz8_4:-KIXzE_1ewXTkRCozFRZ:1727383423) ([raw file](https://github.com/cockroachdb/cockroach/blob/7faa9f846dba58883c29ebb04f83e3366c8900b8/storage/replica.go#L1666)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary>
> Ah, right. The comment should say "not been previously written within this batch", then. But still, where exactly are any reads occurring?
</details>
Every `MVCCPut` does a read of the previous data (unless we're doing a "blind" put). For non-versioned keys like the applied index this read is necessary to keep the MVCC stats up to date.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6861)*
<!-- Sent from Reviewable.io -->