@bdarnell 2016-05-24
Adding a `FreezeFlushes` that lasts for the remainder of the batch seems risky. What if instead we added a variant of `MVCCPut` that takes two `Engines`, one for reads and one for writes? They'd normally be the same, but if we know the value cannot be found in the batch it would pass the underlying engine instead. This would take some more plumbing, but it would avoid the possibility of modifying the behavior of an unrelated operation. 

<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary>
> #### storage/engine: add rocksDBBatchBuilder
> 
> Building batches in Go is significantly faster than performing a cgo call:
> ~55ns/op vs ~400ns/op. The Go batches do not support indexing, so flush to
> a C++ batch whenever a read of the batch is being performed. I experimented
> with adding indexing to the Go side, but all attempts were significantly
> slower than flushing to C++.
> 
> Added an `engine.Batch` interface with a `FreezeFlushes` method which is
> called to avoid flushing a batch in order to write the raft applied index
> and mvcc stats.
> 
> Fixes #6289.
> Fixes #6739.
> 
> ```
> name                            old time/op  new time/op  delta
> TrackChoices1_Cockroach-32       597µs ± 5%   577µs ± 3%   -3.37%  (p=0.000 n=19+20)
> TrackChoices10_Cockroach-32      153µs ± 5%   138µs ± 4%  -10.02%  (p=0.000 n=20+20)
> TrackChoices100_Cockroach-32    96.9µs ± 7%  80.5µs ± 5%  -16.95%  (p=0.000 n=20+19)
> TrackChoices1000_Cockroach-32   92.0µs ± 7%  73.7µs ± 6%  -19.94%  (p=0.000 n=20+19)
> InsertDistinct1_Cockroach-32    2.04ms ± 9%  1.84ms ± 8%   -9.62%  (p=0.000 n=20+20)
> InsertDistinct10_Cockroach-32   1.36ms ± 5%  1.23ms ± 4%  -10.09%  (p=0.000 n=19+20)
> InsertDistinct100_Cockroach-32  2.03ms ±15%  1.67ms ± 7%  -17.92%  (p=0.000 n=20+19)
> KVInsert1_SQL-32                 519µs ± 4%   528µs ± 5%   +1.88%  (p=0.024 n=20+20)
> KVInsert10_SQL-32                926µs ± 6%   906µs ± 3%   -2.13%  (p=0.028 n=20+20)
> KVInsert100_SQL-32              3.89ms ± 8%  3.53ms ± 8%   -9.14%  (p=0.000 n=20+19)
> KVInsert1000_SQL-32             29.4ms ±11%  27.5ms ± 8%   -6.64%  (p=0.000 n=20+20)
> KVInsert10000_SQL-32             356ms ±10%   340ms ±11%   -4.41%  (p=0.013 n=19+19)
> KVUpdate1_SQL-32                 757µs ± 9%   742µs ± 9%     ~     (p=0.253 n=20+20)
> KVUpdate10_SQL-32               1.31ms ± 7%  1.27ms ± 6%   -2.96%  (p=0.001 n=20+19)
> KVUpdate100_SQL-32              5.44ms ± 7%  5.24ms ± 7%   -3.74%  (p=0.006 n=20+20)
> KVUpdate1000_SQL-32             43.5ms ± 6%  42.0ms ± 9%   -3.62%  (p=0.007 n=20+20)
> KVUpdate10000_SQL-32             508ms ±16%   452ms ±14%  -10.97%  (p=0.000 n=20+18)
> ```
> 
> 
> 
</details>


---

Review status: 0 of 14 files reviewed at latest revision, 2 unresolved discussions.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6861#-:-KIZthsKLokSNqHtNd49:1850486356)*
<!-- Sent from Reviewable.io -->